fixed 64 bit warning

+++ platforms/ios/Gearboy/inputmanager.mm
@@ -72,7 +72,7 @@ void InputManager::AddRectRegionEvent(float x, float y, float width, float heigh
 
 void InputManager::ClearRegionEvents(void)
 {
-    int size = m_RegionEventVector.size();
+    int size = (int)m_RegionEventVector.size();
 
     for (int i = 0; i < size; i++)
     {
